Change price fields from double to decimal for financial precision
UX-005: Track.Price changed from double to decimal. UploadTrackRequest
price fields changed from double? to decimal?. WithdrawRequest.Amount
changed to decimal. Removed unnecessary (decimal) casts in catalog/
storefront/creator services. This prevents floating-point rounding
errors in financial calculations (e.g. 19.99 * 100 = 1999 exactly).

Note: A database migration is needed for the Track.Price column change.

+++ src/Cambrian.Application/Services/CatalogService.cs
@@ -99,7 +99,7 @@ public class CatalogService : ICatalogService
             Title = t.Title,
             Description = t.Description,
             Genre = t.Genre ?? "",
-            Price = (decimal)t.Price,
+            Price = t.Price,
             NonExclusivePrice = nonExPrice,
             ExclusivePrice = exPrice,
             CopyrightBuyoutPrice = buyoutPrice,
